Re-Upload of HTML/CSS files and addition of 2 more HTML files.
I have deleted the old html and css files and i have i uploaded the new and correct versions of them.
On the previous files there was a lot of mistakes. So i decided to recreate them and add those to GitHub so we can continue our assignment from the latest and correct files. Also i added 2 more pages; the play page where you add the user name that you want and the PickATreasureHunt where you can pick a treasure hunt. So far after that, the only remaining thing to do is the script and some additions on pages so it gonna work appropriately.